requirements.txt
discord
audioop-lts
python-dotenv
boto3